Fix the contour image version used by the contour-operator
Currently the contour-operator uses the "latest" image to provision the
resources for a Contour resource. However the not every contour image is
guaranteed to work with every operator version. This change fixes the
contour image version used by the operator. When updating the operator
the contour image version will also have to be updated.